[MOO] add OMEditStreamLogger which preserves newlines when logging
