+++ OMCompiler/SimulationRuntime/c/moo/streamlog.h
@@ -1,0 +1,52 @@
+/*
+ * This file is part of OpenModelica.
+ *
+ * Copyright (c) 1998-2025, Linköpings University,
+ * Department of Computer and Information Science,
+ * SE-58183 Linköping, Sweden.
+ *
+ * All rights reserved.
+ *
+ * THIS PROGRAM IS PROVIDED UNDER THE TERMS OF THIS OSMC PUBLIC
+ * LICENSE (OSMC-PL). ANY USE, REPRODUCTION OR DISTRIBUTION OF
+ * THIS PROGRAM CONSTITUTES RECIPIENT'S ACCEPTANCE OF THE OSMC
+ * PUBLIC LICENSE.
+ *
+ * The OpenModelica software and the Open Source Modelica
+ * Consortium (OSMC) Public License (OSMC-PL) are obtained
+ * from Linköpings University, either from the above address,
+ * from the URL: http://www.ida.liu.se/projects/OpenModelica
+ * and in the OpenModelica distribution.
+ *
+ * This program is distributed  WITHOUT ANY WARRANTY; without
+ * even the implied warranty of  MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE, EXCEPT AS EXPRESSLY SET FORTH
+ * IN THE BY RECIPIENT SELECTED SUBSIDIARY LICENSE CONDITIONS
+ * OF OSMC-PL.
+ *
+ * See the full OSMC Public License conditions for more details.
+ *
+ */
+
+#ifndef MOO_OM_STREAMLOG_H
+#define MOO_OM_STREAMLOG_H
+
+#include <base/log.h>
+
+namespace OpenModelica {
+
+class StreamLogger : public Logger {
+public:
+    void log(LogLevel lvl, std::string msg) override;
+};
+
+class OMEditStreamLogger : public Logger {
+public:
+    void log(LogLevel lvl, std::string msg) override;
+};
+
+void create_set_logger();
+
+} // namespace OpenModelica
+
+#endif // MOO_OM_STREAMLOG_H
